chore: change directory
