+++ packages/maskbook/src/extension/options-page/DashboardDialogs/Base.tsx
@@ -13,8 +13,7 @@ import {
 } from '@material-ui/core'
 import { ThemeProvider } from '@material-ui/core/styles'
 import CloseIcon from '@material-ui/icons/Close'
-import { useSnackbar } from 'notistack'
-import { useI18N, extendsTheme, useClassicMaskTheme, useMatchXS } from '../../../utils'
+import { extendsTheme, useClassicMaskTheme, useMatchXS } from '../../../utils'
 
 const useStyles = makeStyles((theme) => ({
     root: {
@@ -270,34 +269,5 @@ export function DashboardDialogWrapper(props: DashboardDialogWrapperProps) {
                 {footer ? <section className={classes.footer}>{footer}</section> : null}
             </DialogContent>
         </ThemeProvider>
-    )
-}
-
-export function useSnackbarCallback<P extends (...args: any[]) => Promise<T>, T>(
-    executor: P,
-    deps: React.DependencyList,
-    onSuccess?: (ret: T) => void,
-    onError?: (err: Error) => void,
-    key?: string,
-    successText?: string,
-) {
-    const { t } = useI18N()
-    const { enqueueSnackbar } = useSnackbar()
-    return useCallback(
-        (...args) =>
-            executor(...args).then(
-                (res) => {
-                    enqueueSnackbar(successText ?? t('done'), { key, variant: 'success', preventDuplicate: true })
-                    onSuccess?.(res)
-                    return res
-                },
-                (err) => {
-                    enqueueSnackbar(`Error: ${err.message || err}`, { key, preventDuplicate: true })
-                    onError?.(err)
-                    throw err
-                },
-            ),
-        // eslint-disable-next-line react-hooks/exhaustive-deps
-        [...deps, enqueueSnackbar, executor, onError, onSuccess, key, successText],
     )
 }
