+++ packages/maskbook/src/extension/options-page/DashboardDialogs/Contact/DeleteConfirm.tsx
@@ -1,10 +1,10 @@
 import { Button } from '@material-ui/core'
 import { UserMinus } from 'react-feather'
-import { useI18N } from '../../../../utils'
+import { useI18N, useSnackbarCallback } from '../../../../utils'
 import Services from '../../../service'
 import { DebounceButton } from '../../DashboardComponents/ActionButton'
 import SpacedButtonGroup from '../../DashboardComponents/SpacedButtonGroup'
-import { DashboardDialogCore, DashboardDialogWrapper, useSnackbarCallback, WrappedDialogProps } from '../Base'
+import { DashboardDialogCore, DashboardDialogWrapper, WrappedDialogProps } from '../Base'
 import type { ContactProps } from './types'
 
 export function DashboardContactDeleteConfirmDialog(
