+++ packages/maskbook/src/extension/options-page/DashboardDialogs/Persona/DeleteConfirm.tsx
@@ -1,11 +1,11 @@
 import { Button } from '@material-ui/core'
 import { UserMinus } from 'react-feather'
-import { useI18N } from '../../../../utils'
+import { useI18N, useSnackbarCallback } from '../../../../utils'
 import { useMyPersonas } from '../../../../components/DataSource/useMyPersonas'
 import Services from '../../../service'
 import { DebounceButton } from '../../DashboardComponents/ActionButton'
 import SpacedButtonGroup from '../../DashboardComponents/SpacedButtonGroup'
-import { DashboardDialogCore, DashboardDialogWrapper, useSnackbarCallback, WrappedDialogProps } from '../Base'
+import { DashboardDialogCore, DashboardDialogWrapper, WrappedDialogProps } from '../Base'
 import type { PersonaProps } from './types'
 
 export function DashboardPersonaDeleteConfirmDialog(props: WrappedDialogProps<PersonaProps>) {
