+++ packages/maskbook/src/extension/options-page/DashboardDialogs/Wallet/AddERC721Token.tsx
@@ -2,12 +2,12 @@ import { TextField } from '@material-ui/core'
 import { useMemo, useState } from 'react'
 import { Octagon as OctagonIcon } from 'react-feather'
 import { EthereumAddress } from 'wallet.ts'
-import { useI18N } from '../../../../utils'
+import { useI18N, useSnackbarCallback } from '../../../../utils'
 import { WalletRPC } from '../../../../plugins/Wallet/messages'
 import { useERC721TokenAssetDetailed } from '../../../../web3/hooks/useERC721TokenAssetDetailed'
 import { useERC721TokenDetailed } from '../../../../web3/hooks/useERC721TokenDetailed'
 import { DebounceButton } from '../../DashboardComponents/ActionButton'
-import { DashboardDialogCore, DashboardDialogWrapper, useSnackbarCallback, WrappedDialogProps } from '../Base'
+import { DashboardDialogCore, DashboardDialogWrapper, WrappedDialogProps } from '../Base'
 import type { WalletProps } from './types'
 
 export function DashboardWalletAddERC721TokenDialog(props: WrappedDialogProps<WalletProps>) {
