+++ packages/maskbook/src/extension/options-page/DashboardDialogs/Wallet/DeleteConfirm.tsx
@@ -1,10 +1,10 @@
 import { Button } from '@material-ui/core'
 import { CreditCard as CreditCardIcon } from 'react-feather'
-import { useI18N } from '../../../../utils'
+import { useI18N, useSnackbarCallback } from '../../../../utils'
 import { WalletRPC } from '../../../../plugins/Wallet/messages'
 import { DebounceButton } from '../../DashboardComponents/ActionButton'
 import SpacedButtonGroup from '../../DashboardComponents/SpacedButtonGroup'
-import { DashboardDialogCore, DashboardDialogWrapper, useSnackbarCallback, WrappedDialogProps } from '../Base'
+import { DashboardDialogCore, DashboardDialogWrapper, WrappedDialogProps } from '../Base'
 import type { WalletProps } from './types'
 
 export function DashboardWalletDeleteConfirmDialog(props: WrappedDialogProps<WalletProps>) {
