+++ packages/maskbook/src/plugins/Snapshot/UI/VotingCard.tsx
@@ -2,7 +2,7 @@ import { useContext, useState, useEffect } from 'react'
 import classNames from 'classnames'
 import { Button } from '@material-ui/core'
 import { makeStyles } from '@material-ui/core'
-import { useI18N } from '../../../utils'
+import { useI18N, useSnackbarCallback } from '../../../utils'
 import { SnapshotContext } from '../context'
 import { useAccount } from '../../../web3/hooks/useAccount'
 import { PluginSnapshotRPC } from '../messages'
@@ -11,7 +11,6 @@ import { useProposal } from '../hooks/useProposal'
 import { usePower } from '../hooks/usePower'
 import { EthereumWalletConnectedBoundary } from '../../../web3/UI/EthereumWalletConnectedBoundary'
 import { VoteConfirmDialog } from './VoteConfirmDialog'
-import { useSnackbarCallback } from '../../../extension/options-page/DashboardDialogs/Base'
 import { useRetry } from '../hooks/useRetry'
 
 const useStyles = makeStyles((theme) => {
