+++ packages/maskbook/src/plugins/Wallet/UI/WalletStatusDialog.tsx
@@ -5,7 +5,7 @@ import ErrorIcon from '@material-ui/icons/Error'
 import { Button, DialogActions, DialogContent, Link, makeStyles, Typography } from '@material-ui/core'
 import { InjectedDialog } from '../../../components/shared/InjectedDialog'
 import { ProviderIcon } from '../../../components/shared/ProviderIcon'
-import { useSnackbarCallback } from '../../../extension/options-page/DashboardDialogs/Base'
+import { useSnackbarCallback } from '../../../utils'
 import Services from '../../../extension/service'
 import { useRemoteControlledDialog, useValueRef, useI18N } from '../../../utils'
 import { useChainId, useChainIdValid } from '../../../web3/hooks/useChainId'
