+++ packages/maskbook/src/utils/hooks/index.ts
@@ -13,3 +13,4 @@ export * from './useSettingSwitcher'
 export * from './useSuspense'
 export * from './useValueRef'
 export * from './useVideoDevices'
+export * from './useSnackbarCallback'
