+++ packages/maskbook/src/utils/hooks/useSnackbarCallback.tsx
@@ -1,0 +1,32 @@
+import { useCallback } from 'react'
+import { useSnackbar } from 'notistack'
+import { useI18N } from '../index'
+
+export function useSnackbarCallback<P extends (...args: any[]) => Promise<T>, T>(
+    executor: P,
+    deps: React.DependencyList,
+    onSuccess?: (ret: T) => void,
+    onError?: (err: Error) => void,
+    key?: string,
+    successText?: string,
+) {
+    const { t } = useI18N()
+    const { enqueueSnackbar } = useSnackbar()
+    return useCallback(
+        (...args) =>
+            executor(...args).then(
+                (res) => {
+                    enqueueSnackbar(successText ?? t('done'), { key, variant: 'success', preventDuplicate: true })
+                    onSuccess?.(res)
+                    return res
+                },
+                (err) => {
+                    enqueueSnackbar(`Error: ${err.message || err}`, { key, preventDuplicate: true })
+                    onError?.(err)
+                    throw err
+                },
+            ),
+        // eslint-disable-next-line react-hooks/exhaustive-deps
+        [...deps, enqueueSnackbar, executor, onError, onSuccess, key, successText],
+    )
+}
